+++ packages/component/src/SpeechServices.js
@@ -4,15 +4,7 @@ import createSpeechRecognitionPonyfill, { createSpeechRecognitionPonyfillFromRec
 import createSpeechSynthesisPonyfill from './SpeechServices/TextToSpeech';
 import fetchAuthorizationToken from './SpeechServices/fetchAuthorizationToken';
 
-let shouldWarnOnSubscriptionKey = true;
-
 export default function createSpeechServicesPonyfill(options = {}, ...args) {
-  if (shouldWarnOnSubscriptionKey && options.subscriptionKey) {
-    console.warn('web-speech-cognitive-services: In production environment, subscription key should not be used, authorization token should be used instead.');
-
-    shouldWarnOnSubscriptionKey = false;
-  }
-
   const ponyfill = {
     ...createSpeechRecognitionPonyfill(options, ...args),
     ...createSpeechSynthesisPonyfill(options, ...args),