+++ packages/component/src/SpeechServices/resolveFunctionOrReturnValue.js
@@ -1,0 +1,3 @@
+export default async function resolveFunctionOrReturnValue(fnOrValue) {
+  return await (typeof fnOrValue === 'function' ? fnOrValue() : fnOrValue);
+}